Another turn

+++ scb/Table.php
@@ -4,15 +4,17 @@
 class scbTable {
 	protected $name;
 	protected $columns;
+	protected $upgrade_method;
 
-	function __construct($name, $file, $columns) {
+	function __construct( $name, $file, $columns, $upgrade_method = 'dbDelta' ) {
 		global $wpdb;
 
 		$this->name = $wpdb->$name = $wpdb->prefix . $name;
 		$this->columns = $columns;
+		$this->upgrade_method = $upgrade_method;
 
-		register_activation_hook($file, array($this, 'install'));
-		scbUtil::add_uninstall_hook($file, array($this, 'uninstall'));
+		scbUtil::add_activation_hook( $file, array( $this, 'install' ) );
+		scbUtil::add_uninstall_hook( $file, array( $this, 'uninstall' ) );
 	}
 
 	function install() {
@@ -20,21 +22,28 @@ class scbTable {
 
 		$charset_collate = '';
 		if ( $wpdb->has_cap( 'collation' ) ) {
-			if ( ! empty($wpdb->charset) )
+			if ( ! empty( $wpdb->charset ) )
 				$charset_collate = "DEFAULT CHARACTER SET $wpdb->charset";
-			if ( ! empty($wpdb->collate) )
+			if ( ! empty( $wpdb->collate ) )
 				$charset_collate .= " COLLATE $wpdb->collate";
 		}
 
-		require_once ABSPATH . 'wp-admin/includes/upgrade.php';
+		if ( 'dbDelta' == $this->upgrade_method ) {
+			require_once ABSPATH . 'wp-admin/includes/upgrade.php';
+			dbDelta( "CREATE TABLE $this->name ( $this->columns ) $charset_collate" );		
+			return;
+		}
 
-		dbDelta("CREATE TABLE $this->name ($this->columns) $charset_collate;");
+		if ( 'delete_first' == $this->upgrade_method )
+			$wpdb->query( "DROP TABLE IF EXISTS $this->name;" );
+
+		$wpdb->query( "CREATE TABLE IF NOT EXISTS $this->name ( $this->columns ) $charset_collate;" );
 	}
 
 	function uninstall() {
 		global $wpdb;
 
-		$wpdb->query("DROP TABLE IF EXISTS $this->name");
+		$wpdb->query( "DROP TABLE IF EXISTS $this->name" );
 	}
 }
